+++ apps/fyler/src/fyler_utils.erl
@@ -19,7 +19,7 @@ stats_to_proplist(#job_stats{status=Status,
   time_spent = Time,
   error_msg = Error,
   task_type = Type}) ->
-  [
+  {[
   {status, Status}, 
   {download_time, DTime},
   {upload_time, UTime},
@@ -27,20 +27,20 @@ stats_to_proplist(#job_stats{status=Status,
   {file_path, Path},
   {time_spent, Time},
   {task_type, Type},
-  {error, error_to_s(Error)}].
+  {error, error_to_s(Error)}]}.
 
 
 current_task_to_proplist(#current_task{id=Id, status=Status, url=Url, type = Type, pool = Pool}) ->
-  [
+  {[
     {id, Id},
     {status, Status},
     {url, Url},
     {type, Type},
     {pool,Pool}
-  ].
+  ]}.
 
 pool_to_proplist(#pool{node = Node, category = Type, enabled = Enabled, active_tasks_num = Active, total_tasks = Total}) ->
-  [{node, Node}, {category, Type}, {enabled, Enabled}, {active_tasks, Active}, {total, Total}]. 
+  {[{node, atom_to_binary(Node,utf8)}, {category, Type}, {enabled, Enabled}, {active_tasks, Active}, {total, Total}]}. 
 
 
 %% @doc
@@ -87,7 +87,7 @@ pgtime_to_string({{Year,Month,Day},{Hour,Minute,Seconds}}) ->
     pad(Minute)++":"++
     pad(trunc(Seconds));
 
-pgtime_to_string(_F) -> {error,badformat,_F}.
+pgtime_to_string(_F) -> "".
 
 
 pad(N) when N < 10 -> "0"++integer_to_list(N);